+++ environments/gridworld687.py
@@ -25,8 +25,7 @@ class Gridworld687(Environment):
         self._horizonLength = horizonLength
         self.numStates = size ** 2
         self.numActions = 4
-        self.observation_space = spaces.Box(low=np.zeros(self.numActions), high=np.ones(self.numActions),
-                                            dtype=np.float32)
+        self.observation_space = spaces.Box(low=np.zeros(self.numActions), high=np.ones(self.numActions))
 
         self.action_space = spaces.Discrete(self.numActions)
         self.x = 0
